replace use of pool::details::singletion by serialization::singleton
The first one has dissapear in recent boost version so :)
boost::serialization::singletion is part of the serialization interface,
so we can hope that it will stay here for a long time.

XXX The patch still depends of a detail part of the implementation, so
it may breaks on some boost version.